% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_drugs_as_nodes}
\alias{add_drugs_as_nodes}
\title{Add Drugs as Nodes}
\usage{
add_drugs_as_nodes(W_mat, drug_target_interactions, edge_weight = 1000)
}
\arguments{
\item{W_mat}{adjacency matrix for the chosen PIN}

\item{drug_target_interactions}{data frame containing (processed) drugs and target genes}

\item{edge_weight}{edge weight for drug-target gene interaction (default = 1000)}
}
\value{
adjacency matrix with the drugs added as nodes
}
\description{
Add Drugs as Nodes
}
