% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{adj_list2mat}
\alias{adj_list2mat}
\title{Turn Adjacency List into Adjacency Matrix}
\usage{
adj_list2mat(adj_list)
}
\arguments{
\item{adj_list}{Adjacency list}
}
\value{
Adjacency matrix
}
\description{
Turn Adjacency List into Adjacency Matrix
}
