% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert2alias}
\alias{convert2alias}
\title{Convert Input Gene Symbols to Alias}
\usage{
convert2alias(input_genes, target_genes)
}
\arguments{
\item{input_genes}{vector of input genes}

\item{target_genes}{vector of target genes}
}
\value{
vector of converted gene symbols (if any alias in target genes)
}
\description{
Convert Input Gene Symbols to Alias
}
