% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DGIdb_interactions_df}
\alias{DGIdb_interactions_df}
\title{DGIdb Interactions Expert-curated Sources}
\format{
a data frame containing 11323 rows and 2 variables:
\describe{
\item{drug_name}{Drug name}
\item{gene_name}{HGNC gene symbol for the interacting gene}
}
}
\usage{
DGIdb_interactions_df
}
\description{
Data frame containing drug-gene interactions from expert-curated sources
(CancerCommons, CGI, ChemblInteractions, CIViC, ClearityFoundationBiomarkers,
ClearityFoundationClinicalTrial, COSMIC, DoCM, MyCancerGenome,
MyCancerGenomeClinicalTrial, TALC, TdgClinicalTrial, TEND)
from DGIdb.
}
\keyword{datasets}
