% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_driveR_res}
\alias{example_driveR_res}
\title{Example driveR Result}
\format{
a data frame containing 106 rows and 3 variables:
\describe{
\item{gene_symbol}{HGNC gene symbol}
\item{driverness_prob}{'driverness' probability}
\item{prediction}{driveR's prediction for whether the gene is a 'driver' or 'non-driver'}
}
}
\usage{
example_driveR_res
}
\description{
Data frame containing 'driveR' results for a lung adenocarcinoma case.
}
\keyword{datasets}
