% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scores_dist}
\alias{example_scores_dist}
\title{Example PANACEA "distance-based" Method Result}
\format{
named vector of 1423 values
}
\usage{
example_scores_dist
}
\description{
Vector containing 'PANACEA' "distance-based" results for a lung adenocarcinoma case.
Names are drug names, values are scores
}
\keyword{datasets}
