% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scores_RWR}
\alias{example_scores_RWR}
\title{Example PANACEA "RWR" Method Result}
\format{
named vector of 1423 values
}
\usage{
example_scores_RWR
}
\description{
Vector containing 'PANACEA' "RWR" results for a lung adenocarcinoma case.
Names are drug names, values are scores
}
\keyword{datasets}
