% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapl_norm.R
\name{Laplacian.norm}
\alias{Laplacian.norm}
\title{Graph Laplacian Normalization}
\usage{
Laplacian.norm(W)
}
\arguments{
\item{W}{square symmetric adjacency matrix}
}
\value{
normalized adjacency matrix
}
\description{
Graph Laplacian Normalization
}
