% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process_drug_target_interactions}
\alias{process_drug_target_interactions}
\title{Process Drug-Target Interactions}
\usage{
process_drug_target_interactions(
  drug_target_interactions,
  PIN_genes,
  drug_name_col = "drug_name",
  target_col = "gene_name"
)
}
\arguments{
\item{drug_target_interactions}{data frame containing drugs and target genes}

\item{PIN_genes}{gene symbols for the chosen PIN}

\item{drug_name_col}{name of the column containing drug names (default = "drug_name")}

\item{target_col}{name of the column containing drug targets (default = "converted_target_gene")}
}
\value{
processed drug-target interactions. Processing involves converting
symbols missing in the PIN, merging drugs that have the same target gene(s)
}
\description{
Process Drug-Target Interactions
}
