% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRING_adj_df}
\alias{STRING_adj_df}
\title{Adjacency List for STRING v11.5 - High Confidence Interactions}
\format{
a data frame with 887797 rows and 3 variables:
\describe{
\item{protein1}{Interactor 1}
\item{protein2}{Interactor 2}
\item{value}{edge weight(combined score)}
}
}
\usage{
STRING_adj_df
}
\description{
Data frame of adjacency list for STRING v11.5 interactions with combined
score > 700 (high confidence)
}
\keyword{datasets}
