% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_W_mat}
\alias{toy_W_mat}
\title{Toy Adjacency Matrix (for examples)}
\format{
matrix of 84 rows and 84 columns
}
\usage{
toy_W_mat
}
\description{
Symmetric matrix containing example adjacency data
}
\keyword{datasets}
